+++ src/main/java/kitchenpos/menu/menugroup/application/MenuGroupService.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu.menugroup.application;
 
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.menugroup.domain.MenuGroup;
+import kitchenpos.menu.menugroup.domain.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
