+++ src/main/java/kitchenpos/order/delivery/application/OrderService.java
@@ -4,8 +4,8 @@ import kitchenpos.domain.*;
 import kitchenpos.order.delivery.domain.*;
 import kitchenpos.order.eatin.domain.*;
 import kitchenpos.order.delivery.infra.KitchenridersClient;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.menu.domain.Menu;
+import kitchenpos.menu.menu.domain.MenuRepository;
 import kitchenpos.order.domain.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
