+++ src/main/java/kitchenpos/order/delivery/domain/OrderLineItem.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.delivery.domain;
 
-import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.menu.domain.Menu;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
